+++ src/sql/Cargo.toml
@@ -15,6 +15,7 @@ datafusion-sql.workspace = true
 datatypes = { workspace = true }
 hex = "0.4"
 itertools.workspace = true
+lazy_static.workspace = true
 once_cell.workspace = true
 regex.workspace = true
 snafu = { version = "0.7", features = ["backtraces"] }
